% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.rsfit}
\alias{predict.rsfit}
\title{predict.rsfit get the predicted relative contrast function}
\usage{
\method{predict}{rsfit}(fit, newx, lossType = "logistic", type = "transformed")
}
\arguments{
\item{fit}{the list fit in the output of rsfit.}

\item{newx}{input matrix of dimension nobs x nvars. Each raw is a new observation, each column is a covariate}
}
\value{
A Array
\describe{
A array includes all predictions.
}
}
\description{
predict.rsfit get the predicted relative contrast function
}
\references{
Muxuan Liang, Menggang Yu (2022). Relative Contrast Estimation and Inference for Treatment Recommendation.
}
\author{
Muxuan Liang
}
