% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfit.R
\name{rsfit}
\alias{rsfit}
\title{Estimation and inference for relative contrast function under a monotonic single-index model assumption}
\usage{
rsfit(
  covariate,
  response,
  treatment,
  splitIndex = NULL,
  propensityModel = "glmnet",
  estimatedPropensity = NULL,
  outcomeModel = "kernel",
  estimatedOutcome = NULL,
  lossType = "logistic",
  weights = NULL,
  tol = 0.001,
  propensityFormula = NULL,
  outcomeFormula = NULL,
  parallel = FALSE,
  constraint = TRUE,
  boundaryPoint = c(-1, 1),
  efficient = TRUE,
  local = TRUE
)
}
\arguments{
\item{covariate}{input matrix of dimension nobs x nvars. Each raw is a observation, each column is a covariate}

\item{response}{numeric response}

\item{treatment}{is a vector of binary value representing two treatment, 0 or 1.}

\item{propensityModel}{Similar to outcomeModel.}

\item{estimatedPropensity}{estimated propensity score p(trt=1|X). This should be used only when the propensity is estimated by a parametric model.}

\item{outcomeModel}{this selects method to estimate the outcome when estimatedOutcome=NULL. Options include lm, glmnet, kernel, and gam. If lm is used, user also need to input the outcomeFormula like y~x used in lm. By default, kernel regression is selected.}

\item{estimatedOutcome}{estimated outcome model E(y|trt, X). This should be used only when the outcome is estimated by a parametric model. It should be a list (list(control=..., treatment=...)).}

\item{boundaryPoint}{please specify appropriate range for the single-index. Default is [-1,1].}
}
\value{
A list
\describe{
\item{betaAN}{The one-step updated coefficients estimates}
\item{sigmaAN}{The estimated sd of the estimated coefficients}
\item{betaAN.renorm}{The one-step updated coefficients estimates after re-normalized such that \|\beta\|_2=1}
\item{sigmaAN.renorm}{The estimated sd of the normalized estimates}
\item{fit}{A list used for predict.rsfit which predicts the relative contrast}
}
}
\description{
Estimation and inference for relative contrast function under a monotonic single-index model assumption
}
\references{
Muxuan Liang, Menggang Yu (2022). Relative Contrast Estimation and Inference for Treatment Recommendation.
}
\author{
Muxuan Liang
}
